replacing old text_encoder scripts with tf version
